Split Excel autofit text helpers
